refactor(cli): split main.rs into rust/src/cli/* per subcommand (#267 F6)
`main.rs` was 574 LOC mixing clap derive defs, three list-voices helpers,
the 133-LOC `run_say` body, and the install warm-up logic. After this PR
it's 207 LOC and contains only the `Cli` parser + `Commands` enum +
`main()` dispatch table — matching the audit's "thin parse + dispatch"
shape and mirroring the TS-side `src/cli/*` layout.

New module structure:

- `rust/src/cli/mod.rs` — façade `pub mod` declarations only.
- `rust/src/cli/say.rs` — `SayArgs`, `run_say` (→ `pub fn run`),
  `list_kokoro_voices`, `list_vosk_ru_voices`, `exit_code_for_tts_err`,
  and `resolve_output_format` (still `pub(crate)`; `say_loop.rs` calls
  through to it via the new path).
- `rust/src/cli/install.rs` — `pub fn run(no_cache, tts, vad, diarize,
  no_warmup)` carrying the model-download + ASR-backend warm-up body.
  Conditional fields stay `#[cfg(feature = "…")]`-gated as before.
- `rust/src/cli/transcribe.rs` — `pub fn run(audio_path, json, vad,
  no_vad, speakers)`.
- `rust/src/cli/detect_lang.rs` — `pub fn run(audio_path)`.
- `rust/src/cli/detect_text_lang.rs` — `pub fn run(text)`.

`main()` keeps:

- `debug::init()` BEFORE `Cli::parse()` (anchors the dtrace! `+Nms`
  timeline; Greptile P2 on #293 fixed exactly this ordering).
- `--capabilities-json` early-return (no subcommand reach).
- A 5-arm `match cli.command` that hands every subcommand off to
  `cli::<sub>::run(...)`.
- `Commands` enum + per-arm clap attributes (kept inline so all clap
  derive definitions live next to the parser).

`say_loop.rs` updated to call `crate::cli::say::resolve_output_format`
instead of the previous `crate::resolve_output_format`. Same function,
new path.

Verification:

- `cargo nextest run --features tts` → **449/449 tests pass**, 0 skipped.
- `cargo clippy --all-targets --features tts -- -D warnings` clean.
- `cargo fmt --check` clean (rustfmt rewrapped `resolve_output_format`
  call in say_loop.rs across 5 lines + `cli/transcribe.rs::run`
  parameters onto one line — both stylistic, non-semantic).
- `cargo build --features tts` clean.
- `cargo check --features coreml,tts --no-default-features` fails on
  the same local Swift-toolchain issue inside `fluidaudio-rs` build.rs
  that reproduces on `main` without these changes (CI's macos-14 +
  pinned Xcode 16.2 handles it cleanly via rust-test.yml).

No behaviour change at the CLI surface or in the public Rust API.
All subcommand outputs, exit codes, and clap help text identical.

Refs #267
